+++ examples/bench/Probe.hs
@@ -8,7 +8,7 @@
 module Main (main) where
 
 import Control.Monad (forM_)
-import qualified Data.Text as T
+import qualified Data.ByteString as BS
 import GHC.Clock (getMonotonicTime)
 import GHC.IO (evaluate)
 import JShark (ClosedEffect, renderJS)
@@ -38,7 +38,7 @@ runOne :: Int -> IO ()
 runOne n = do
   start <- getMonotonicTime
   let
-    bytes = T.length (renderJS (effectfulAST (progN n)))
+    bytes = BS.length (renderJS (effectfulAST (progN n)))
   bytes `seq` pure ()
   end <- getMonotonicTime
   evaluate bytes